+++ dependencies/pom.xml
@@ -117,7 +117,7 @@
         <version.lib.yasson>1.0.6</version.lib.yasson>
         <version.lib.zipkin>2.12.5</version.lib.zipkin>
         <version.lib.zipkin.sender-urlconnection>2.12.0</version.lib.zipkin.sender-urlconnection>
-        <version.lib.kafka>2.4.1</version.lib.kafka>
+        <version.lib.kafka>2.3.1</version.lib.kafka>
         <version.lib.kafka-junit5>3.2.1</version.lib.kafka-junit5>
         <version.lib.reactive-streams-tck>1.0.3</version.lib.reactive-streams-tck>
     </properties>
