+++ messaging/connectors/kafka/src/main/java/io/helidon/messaging/connectors/kafka/ResourceConnector.java
@@ -1,0 +1,34 @@
+/*
+ * Copyright (c) 2020 Oracle and/or its affiliates. All rights reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+
+package io.helidon.messaging.connectors.kafka;
+
+/**
+ * A ResourceConnector is a source or destination of data that can be closed.
+ * The close method is invoked to release resources that the object is
+ * holding
+ *
+ */
+public interface ResourceConnector {
+
+    /**
+     * Closes this connector and releases any system resources associated
+     * with it.
+     */
+    void close();
+
+}
