+++ messaging/connectors/pom.xml
@@ -1,0 +1,40 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!--
+  ~ Copyright (c)  2020 Oracle and/or its affiliates. All rights reserved.
+  ~
+  ~ Licensed under the Apache License, Version 2.0 (the "License");
+  ~ you may not use this file except in compliance with the License.
+  ~ You may obtain a copy of the License at
+  ~
+  ~     http://www.apache.org/licenses/LICENSE-2.0
+  ~
+  ~ Unless required by applicable law or agreed to in writing, software
+  ~ distributed under the License is distributed on an "AS IS" BASIS,
+  ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+  ~ See the License for the specific language governing permissions and
+  ~ limitations under the License.
+  ~
+  -->
+
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <modelVersion>4.0.0</modelVersion>
+    <parent>
+        <groupId>io.helidon.messaging</groupId>
+        <artifactId>helidon-messaging-project</artifactId>
+        <version>2.0.0-SNAPSHOT</version>
+    </parent>
+    <packaging>pom</packaging>
+    <groupId>io.helidon.messaging.connectors</groupId>
+    <artifactId>helidon-messaging-connectors</artifactId>
+    <name>Helidon Messaging Connectors</name>
+
+    <description>
+        Helidon Messaging Connectors implementation
+    </description>
+
+    <modules>
+        <module>kafka</module>
+    </modules>
+</project>
