+++ messaging/pom.xml
@@ -27,32 +27,11 @@
         <version>2.0.0-SNAPSHOT</version>
     </parent>
     <packaging>pom</packaging>
-    <groupId>io.helidon.microprofile</groupId>
-    <artifactId>helidon-microprofile-project</artifactId>
-    <name>Helidon Microprofile</name>
+    <groupId>io.helidon.messaging</groupId>
+    <artifactId>helidon-messaging-project</artifactId>
+    <name>Helidon Messaging</name>
 
     <modules>
-        <module>config</module>
-        <module>health</module>
-        <module>metrics</module>
-        <module>oidc</module>
-        <module>server</module>
-        <module>security</module>
-        <module>fault-tolerance</module>
-        <module>bundles</module>
-        <module>tests</module>
-        <module>jwt-auth</module>
-        <module>tracing</module>
-        <module>openapi</module>
-        <module>rest-client</module>
-        <module>access-log</module>
-        <module>grpc</module>
-        <module>cdi</module>
-        <module>weld</module>
-        <module>websocket</module>
-        <module>reactive-streams</module>
-        <module>messaging</module>
-        <module>cors</module>
         <module>connectors</module>
     </modules>
 </project>
