--- microprofile/connectors/kafka/src/main/java/io/helidon/microprofile/connectors/kafka/PartitionsAssignedLatch.java
@@ -1,48 +1,0 @@
-/*
- * Copyright (c) 2020 Oracle and/or its affiliates. All rights reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package io.helidon.microprofile.connectors.kafka;
-
-import java.util.Collection;
-import java.util.concurrent.CountDownLatch;
-import java.util.logging.Logger;
-
-import org.apache.kafka.clients.consumer.ConsumerRebalanceListener;
-import org.apache.kafka.common.TopicPartition;
-
-/**
- * Waiting latch for partition assigment, after that is consumer ready to receive.
- */
-class PartitionsAssignedLatch extends CountDownLatch implements ConsumerRebalanceListener {
-
-    private static final Logger LOGGER = Logger.getLogger(PartitionsAssignedLatch.class.getName());
-
-    PartitionsAssignedLatch() {
-        super(1);
-    }
-
-    @Override
-    public void onPartitionsRevoked(Collection<TopicPartition> partitions) {
-        LOGGER.fine("Partitions revoked: " + partitions);
-        // Do nothing
-    }
-
-    @Override
-    public void onPartitionsAssigned(Collection<TopicPartition> partitions) {
-        LOGGER.fine("Partitions assigned: " + partitions);
-        this.countDown();
-    }
-}
