+++ pom.xml
@@ -169,6 +169,7 @@
         <module>webclient</module>
         <module>integrations</module>
         <module>dbclient</module>
+        <module>messaging</module>
     </modules>
