+++ tests/integration/kafka/src/test/resources/logging.properties
